requirements.txt
Faker==23.1.0
psycopg2==2.9.9
robotframework==7.0
robotframework-appiumlibrary==2.0.0
robotframework-browser==18.1.0
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.6
python_dotenv=1.0.1
